requirements.txt
absl-py==2.3.1
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
deepspeed==0.17.2
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.47.7
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.6
openai==1.93.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rouge_score==0.1.2
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sentry-sdk==2.32.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sympy==1.13.1
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.49.0
trl==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
wandb==0.18.3
xxhash==3.5.0
yarl==1.20.1
